--- src/lib/Scenes/Favorites/Components/Artists/Relay/FavoriteArtists.tsx
@@ -1,22 +1,0 @@
-import React from "react"
-import { graphql, QueryRenderer } from "react-relay"
-
-import { FavoriteArtistsQuery } from "__generated__/FavoriteArtistsQuery.graphql"
-import { defaultEnvironment } from "lib/relay/createEnvironment"
-
-export default ({ render }: any /* STRICTNESS_MIGRATION */) => {
-  return (
-    <QueryRenderer<FavoriteArtistsQuery>
-      environment={defaultEnvironment}
-      query={graphql`
-        query FavoriteArtistsQuery {
-          me {
-            ...Artists_me
-          }
-        }
-      `}
-      variables={{ count: 10 }}
-      render={render}
-    />
-  )
-}
