--- src/lib/Scenes/Favorites/Components/Artists/__tests__/index-tests.tsx
@@ -1,5 +1,0 @@
-import FavoriteArtists from "../index"
-
-describe(FavoriteArtists, () => {
-  it.todo("No tests yet, it's just placeholders so far.")
-})
