--- src/lib/Scenes/Favorites/Components/Artworks/Relay/FavoriteArtworks.tsx
@@ -1,24 +1,0 @@
-import React from "react"
-import { graphql, QueryRenderer } from "react-relay"
-
-import { FavoriteArtworksQuery } from "__generated__/FavoriteArtworksQuery.graphql"
-import { defaultEnvironment } from "lib/relay/createEnvironment"
-
-export default ({ render }: any /* STRICTNESS_MIGRATION */) => {
-  return (
-    <QueryRenderer<FavoriteArtworksQuery>
-      environment={defaultEnvironment}
-      query={graphql`
-        query FavoriteArtworksQuery {
-          me {
-            ...Artworks_me
-          }
-        }
-      `}
-      variables={{
-        count: 10,
-      }}
-      render={render}
-    />
-  )
-}
