--- src/lib/Scenes/Favorites/Components/Artworks/__tests__/index-tests.tsx
@@ -1,5 +1,0 @@
-import FavoriteArtworks from "../index"
-
-describe(FavoriteArtworks, () => {
-  it.todo("No tests yet, it's just placeholders so far.")
-})
