--- src/lib/Scenes/Favorites/Components/Categories/Relay/FavoriteCategories.tsx
@@ -1,24 +1,0 @@
-import React from "react"
-import { graphql, QueryRenderer } from "react-relay"
-
-import { FavoriteCategoriesQuery } from "__generated__/FavoriteCategoriesQuery.graphql"
-import { defaultEnvironment } from "lib/relay/createEnvironment"
-
-export default ({ render }: any /* STRICTNESS_MIGRATION */) => {
-  return (
-    <QueryRenderer<FavoriteCategoriesQuery>
-      environment={defaultEnvironment}
-      query={graphql`
-        query FavoriteCategoriesQuery {
-          me {
-            ...Categories_me
-          }
-        }
-      `}
-      variables={{
-        count: 10,
-      }}
-      render={render}
-    />
-  )
-}
