--- src/lib/Scenes/Favorites/Components/Categories/__tests__/index-tests.tsx
@@ -1,5 +1,0 @@
-import FavoriteCategories from "../index"
-
-describe(FavoriteCategories, () => {
-  it.todo("No tests yet, it's just placeholders so far.")
-})
