--- src/lib/Scenes/Favorites/Components/Fairs/Relay/FavoriteFairs.tsx
@@ -1,24 +1,0 @@
-import React from "react"
-import { graphql, QueryRenderer } from "react-relay"
-
-import { FavoriteFairsQuery } from "__generated__/FavoriteFairsQuery.graphql"
-import { defaultEnvironment } from "lib/relay/createEnvironment"
-
-export default ({ render }: any /* STRICTNESS_MIGRATION */) => {
-  return (
-    <QueryRenderer<FavoriteFairsQuery>
-      environment={defaultEnvironment}
-      query={graphql`
-        query FavoriteFairsQuery {
-          me {
-            ...Fairs_me
-          }
-        }
-      `}
-      variables={{
-        count: 10,
-      }}
-      render={render}
-    />
-  )
-}
