--- src/lib/Scenes/Favorites/Components/Fairs/__tests__/index-tests.tsx
@@ -1,5 +1,0 @@
-import FavoriteFairs from "../index"
-
-describe(FavoriteFairs, () => {
-  it.todo("No tests yet, it's just placeholders so far.")
-})
