--- src/lib/Scenes/Favorites/Components/Shows/__tests__/index-tests.tsx
@@ -1,5 +1,0 @@
-import FavoriteShows from "../index"
-
-describe(FavoriteShows, () => {
-  it.todo("No tests yet, it's just placeholders so far.")
-})
