+++ src/lib/Scenes/Favorites/FavoriteArtists.tsx
@@ -1,20 +1,24 @@
 import React from "react"
 import { RefreshControl } from "react-native"
-import { createPaginationContainer, graphql, RelayPaginationProp } from "react-relay"
+import { createPaginationContainer, graphql, QueryRenderer, RelayPaginationProp } from "react-relay"
 
 import { SavedItemRow } from "lib/Components/Lists/SavedItemRow"
 import Spinner from "lib/Components/Spinner"
 import { ZeroState } from "lib/Components/States/ZeroState"
+
 import { PAGE_SIZE } from "lib/data/constants"
 
 import { Spacer } from "@artsy/palette"
-import { Categories_me } from "__generated__/Categories_me.graphql"
+import { Artists_me } from "__generated__/Artists_me.graphql"
+import { FavoriteArtistsQuery } from "__generated__/FavoriteArtistsQuery.graphql"
 import { StickyTabPageFlatList } from "lib/Components/StickyTabPage/StickyTabPageFlatList"
 import { StickyTabPageScrollView } from "lib/Components/StickyTabPage/StickyTabPageScrollView"
+import { defaultEnvironment } from "lib/relay/createEnvironment"
 import { extractNodes } from "lib/utils/extractNodes"
+import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 
 interface Props {
-  me: Categories_me
+  me: Artists_me
   relay: RelayPaginationProp
 }
 
@@ -23,7 +27,7 @@ interface State {
   refreshingFromPull: boolean
 }
 
-export class Categories extends React.Component<Props, State> {
+class Artists extends React.Component<Props, State> {
   state = {
     fetchingMoreData: false,
     refreshingFromPull: false,
@@ -38,7 +42,7 @@ export class Categories extends React.Component<Props, State> {
     this.props.relay.loadMore(PAGE_SIZE, error => {
       if (error) {
         // FIXME: Handle error
-        console.error("Categories/index.tsx", error.message)
+        console.error("Artists/index.tsx", error.message)
       }
       this.setState({ fetchingMoreData: false })
     })
@@ -49,7 +53,7 @@ export class Categories extends React.Component<Props, State> {
     this.props.relay.refetchConnection(PAGE_SIZE, error => {
       if (error) {
         // FIXME: Handle error
-        console.error("Categories/index.tsx #handleRefresh", error.message)
+        console.error("Artists/index.tsx #handleRefresh", error.message)
       }
       this.setState({ refreshingFromPull: false })
     })
@@ -57,19 +61,22 @@ export class Categories extends React.Component<Props, State> {
 
   // @TODO: Implement test on this component https://artsyproduct.atlassian.net/browse/LD-563
   render() {
-    const rows = extractNodes(this.props.me.followsAndSaves?.genes, node => node.gene!).map(gene => ({
-      key: gene.id,
-      content: <SavedItemRow square_image href={gene.href!} image={gene.image!} name={gene.name!} />,
-    }))
+    const rows = extractNodes(this.props.me.followsAndSaves?.artists, node => node.artist!).map(
+      ({ href, image, name, id }) => ({
+        key: id,
+        content: <SavedItemRow href={href!} image={image!} name={name!} size={50} />,
+      })
+    )
 
     if (rows.length === 0) {
       return (
         <StickyTabPageScrollView
+          contentContainerStyle={{ flexGrow: 1 }}
           refreshControl={<RefreshControl refreshing={this.state.refreshingFromPull} onRefresh={this.handleRefresh} />}
         >
           <ZeroState
-            title="You’re not following any categories yet"
-            subtitle="Find a few categories to help improve your artwork recommendations."
+            title="You haven’t followed any artists yet"
+            subtitle="When you’ve found an artist you like, follow them to get updates on new works that become available."
           />
         </StickyTabPageScrollView>
       )
@@ -77,12 +84,12 @@ export class Categories extends React.Component<Props, State> {
 
     return (
       <StickyTabPageFlatList
-        style={{ paddingHorizontal: 0, paddingTop: 15 }}
         data={rows}
-        ItemSeparatorComponent={() => <Spacer mb="5px" />}
         onEndReached={this.loadMore}
         onEndReachedThreshold={0.2}
         refreshControl={<RefreshControl refreshing={this.state.refreshingFromPull} onRefresh={this.handleRefresh} />}
+        style={{ paddingTop: 15, paddingHorizontal: 0 }}
+        ItemSeparatorComponent={() => <Spacer mb="5px" />}
         ListFooterComponent={
           this.state.fetchingMoreData ? <Spinner style={{ marginTop: 20, marginBottom: 20 }} /> : null
         }
@@ -91,21 +98,17 @@ export class Categories extends React.Component<Props, State> {
   }
 }
 
-export default createPaginationContainer(
-  Categories,
+const FavoriteArtistsContainer = createPaginationContainer(
+  Artists,
   {
     me: graphql`
-      fragment Categories_me on Me
+      fragment Artists_me on Me
         @argumentDefinitions(count: { type: "Int", defaultValue: 10 }, cursor: { type: "String" }) {
         followsAndSaves {
-          genes: genesConnection(first: $count, after: $cursor) @connection(key: "Categories_followed_genes") {
-            pageInfo {
-              endCursor
-              hasNextPage
-            }
+          artists: artistsConnection(first: $count, after: $cursor) @connection(key: "Artists_artists") {
             edges {
               node {
-                gene {
+                artist {
                   id
                   name
                   href
@@ -123,8 +126,7 @@ export default createPaginationContainer(
   {
     direction: "forward",
     getConnectionFromProps(props) {
-      // @ts-ignore STRICTNESS_MIGRATION
-      return props.me && props.me.followsAndSaves.genes
+      return props.me.followsAndSaves?.artists
     },
     getFragmentVariables(prevVars, totalCount) {
       return {
@@ -136,11 +138,28 @@ export default createPaginationContainer(
       return pageInfo
     },
     query: graphql`
-      query CategoriesMeQuery($count: Int!, $cursor: String) {
+      query ArtistsMeQuery($count: Int!, $cursor: String) {
         me {
-          ...Categories_me @arguments(count: $count, cursor: $cursor)
+          ...Artists_me @arguments(count: $count, cursor: $cursor)
         }
       }
     `,
   }
 )
+
+export const FavoriteArtistsQueryRenderer = () => {
+  return (
+    <QueryRenderer<FavoriteArtistsQuery>
+      environment={defaultEnvironment}
+      query={graphql`
+        query FavoriteArtistsQuery {
+          me {
+            ...Artists_me
+          }
+        }
+      `}
+      variables={{ count: 10 }}
+      render={renderWithLoadProgress(FavoriteArtistsContainer)}
+    />
+  )
+}
