+++ src/lib/Scenes/Favorites/FavoriteArtworks.tsx
@@ -1,16 +1,20 @@
 import React, { Component } from "react"
 import { RefreshControl } from "react-native"
-import { createPaginationContainer, graphql, RelayPaginationProp } from "react-relay"
+import { createPaginationContainer, graphql, QueryRenderer, RelayPaginationProp } from "react-relay"
 
-import GenericGrid from "lib/Components/ArtworkGrids/GenericGrid"
+import GenericGrid, { GenericGridPlaceholder } from "lib/Components/ArtworkGrids/GenericGrid"
 import { ZeroState } from "lib/Components/States/ZeroState"
 import { PAGE_SIZE } from "lib/data/constants"
 
 import { Button } from "@artsy/palette"
 import { Artworks_me } from "__generated__/Artworks_me.graphql"
+import { FavoriteArtworksQuery } from "__generated__/FavoriteArtworksQuery.graphql"
 import { StickyTabPageScrollView } from "lib/Components/StickyTabPage/StickyTabPageScrollView"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
+import { defaultEnvironment } from "lib/relay/createEnvironment"
 import { extractNodes } from "lib/utils/extractNodes"
+import { renderWithPlaceholder } from "lib/utils/renderWithPlaceholder"
+import { useScreenDimensions } from "lib/utils/useScreenDimensions"
 
 interface Props {
   me: Artworks_me
@@ -102,7 +106,7 @@ export class SavedWorks extends Component<Props, State> {
   }
 }
 
-export default createPaginationContainer(
+const FavoriteArtworksContainer = createPaginationContainer(
   SavedWorks,
   {
     me: graphql`
@@ -156,3 +160,32 @@ export default createPaginationContainer(
     `,
   }
 )
+
+export const FavoriteArtworksQueryRenderer = () => {
+  const screen = useScreenDimensions()
+  return (
+    <QueryRenderer<FavoriteArtworksQuery>
+      environment={defaultEnvironment}
+      query={graphql`
+        query FavoriteArtworksQuery {
+          me {
+            ...Artworks_me
+          }
+        }
+      `}
+      variables={{
+        count: 10,
+      }}
+      render={renderWithPlaceholder({
+        Container: FavoriteArtworksContainer,
+        renderPlaceholder: () => {
+          return (
+            <StickyTabPageScrollView scrollEnabled={false} style={{ paddingTop: 20 }}>
+              <GenericGridPlaceholder width={screen.width - 40} />
+            </StickyTabPageScrollView>
+          )
+        },
+      })}
+    />
+  )
+}
