--- src/lib/Scenes/Favorites/__tests__/index-tests.tsx
@@ -1,5 +1,0 @@
-import Favorites from "../index"
-
-describe(Favorites, () => {
-  it.todo("No tests yet, it's just placeholders so far.")
-})
